Add AI trading team example variants
