+++ docsrc/agents.rst
@@ -24,6 +24,7 @@ BotSpot is strongest when you want the agent workflow plus the managed cloud aro
 
    agents_quickstart
    agents_flows
+   agents_examples
    agents_builtin_tools
    agents_investment_committee
    agents_canonical_demos
