+++ scripts/run_ai_committee_provider_benchmark.py
@@ -30,7 +30,9 @@ os.environ.setdefault("LUMIBOT_DISABLE_BACKTEST_PERFORMANCE_TRACKING", "1")
 from lumibot.backtesting import YahooDataBacktesting
 from lumibot.components.agents.manager import _usage_breakdown
 from lumibot.entities import Asset, TradingFee
-from lumibot.example_strategies.ai_investment_committee import AIInvestmentCommitteeStrategy
+from lumibot.example_strategies.ai_trading_team_bull_bear_large_cap_stocks import (
+    AITradingTeamBullBearLargeCapStocksStrategy,
+)
 
 
 ARTIFACT_ROOT = Path("artifacts") / "ai_committee_provider_benchmarks"
@@ -354,7 +356,7 @@ def _run_one_model(model: str, args: argparse.Namespace, root: Path) -> dict[str
     started = time.perf_counter()
     trading_fee = TradingFee(percent_fee=0.001)
 
-    result, strategy = AIInvestmentCommitteeStrategy.run_backtest(
+    result, strategy = AITradingTeamBullBearLargeCapStocksStrategy.run_backtest(
         datasource_class=YahooDataBacktesting,
         backtesting_start=_parse_date(args.start),
         backtesting_end=_parse_date(args.end),
